+++ test/bit_counter_test.dart
@@ -5,7 +5,7 @@ import 'package:bit_array/bit_array.dart';
 void main() {
   group('BitCounter', () {
     test('set and get values', () {
-      final counter = new BitCounter(128);
+      final counter = BitCounter(128);
       for (int c = 0; c < 128; c++) {
         expect(counter[c], 0);
         for (int i = 0; i < 1024; i++) {
@@ -21,7 +21,7 @@ void main() {
     });
 
     test('increment value', () {
-      final counter = new BitCounter(128);
+      final counter = BitCounter(128);
       for (int c = 0; c < 128; c++) {
         for (int i = 1; i < 1024; i++) {
           counter.increment(c);
@@ -36,12 +36,24 @@ void main() {
       expect(counter.bitLength, 10);
     });
 
+    test('increment value with shift', () {
+      final counter = BitCounter(128);
+      counter.increment(10, shiftLeft: 3);
+      counter.increment(10, shiftLeft: 2);
+      expect(counter[10], 12);
+      counter.increment(10);
+      expect(counter[10], 13);
+      counter.increment(10, shiftLeft: 2);
+      expect(counter[10], 17);
+      expect(counter.bitLength, 5);
+    });
+
     test('simple bit arrays', () {
-      final counter = new BitCounter(128);
-      counter.addBitArray(new BitArray(128)..setBits([0, 3]));
-      counter.addBitArray(new BitArray(128)..setBits([0, 2]));
-      counter.addBitArray(new BitArray(128)..setBits([0, 1, 3]));
-      counter.addBitArray(new BitArray(128)..setBits([63]));
+      final counter = BitCounter(128);
+      counter.addBitArray(BitArray(128)..setBits([0, 3]));
+      counter.addBitArray(BitArray(128)..setBits([0, 2]));
+      counter.addBitArray(BitArray(128)..setBits([0, 1, 3]));
+      counter.addBitArray(BitArray(128)..setBits([63]));
       expect(counter[0], 3);
       expect(counter[1], 1);
       expect(counter[2], 1);
@@ -53,9 +65,9 @@ void main() {
     });
 
     test('complex bit arrays', () {
-      final counter = new BitCounter(1024);
+      final counter = BitCounter(1024);
       for (int i = 1; i < 1024; i++) {
-        final array = new BitArray(1024)..setBit(i);
+        final array = BitArray(1024)..setBit(i);
         for (int j = 0; j < i; j++) {
           counter.addBitArray(array);
         }
@@ -67,11 +79,11 @@ void main() {
     });
 
     test('bit sets', () {
-      final counter = new BitCounter(0);
-      counter.addBitSet(new ListSet.fromSorted([0, 2, 5, 2000]));
+      final counter = BitCounter(0);
+      counter.addBitSet(ListSet.fromSorted([0, 2, 5, 2000]));
       expect(counter[2000], 1);
-      counter.addBitSet(new ListSet.fromSorted([2, 2000]));
-      counter.addBitSet(new RangeSet.fromSortedRangeLength([0, 2, 1999, 1]));
+      counter.addBitSet(ListSet.fromSorted([2, 2000]));
+      counter.addBitSet(RangeSet.fromSortedRangeLength([0, 2, 1999, 1]));
       expect(counter[0], 2);
       expect(counter[1], 1);
       expect(counter[2], 3);
@@ -83,6 +95,52 @@ void main() {
       expect(counter[1999], 1);
       expect(counter[2000], 3);
       expect(counter[2001], 0);
+    });
+
+    test('bit set with shift', () {
+      final counter = BitCounter(128);
+      counter.addBitSet(ListSet.fromSorted([0, 2, 5]), shiftLeft: 3);
+      expect(counter[2], 8);
+      counter.addBitSet(ListSet.fromSorted([2, 2000]), shiftLeft: 2);
+      expect(counter[2], 12);
+      expect(counter[2000], 4);
+    });
+
+    test('multiply', () {
+      final c1 = BitCounter(128);
+      for (int i = 0; i < 128; i++) {
+        c1[i] = i;
+      }
+      final c3 = c1.multiply(5);
+      expect(c3[0], 0);
+      expect(c3[10], 50);
+      expect(c3[98], 490);
+      for (int i = 0; i < 128; i++) {
+        final exp = i * 5;
+        expect('$i-${c3[i]}', '$i-$exp');
+      }
+      expect(c1.bitLength, 7);
+      expect(c3.bitLength, 10);
+    });
+
+    test('multiplyWithCounter', () {
+      final c1 = BitCounter(128);
+      final c2 = BitCounter(128);
+      for (int i = 0; i < 128; i++) {
+        c1[i] = i;
+        c2[i] = (i % 17) + (i % 3);
+      }
+      final c3 = c1.multiplyWithCounter(c2);
+      expect(c3[0], 0);
+      expect(c3[10], 110);
+      expect(c3[98], 1470);
+      for (int i = 0; i < 128; i++) {
+        final exp = i * ((i % 17) + (i % 3));
+        expect('$i-${c3[i]}', '$i-$exp');
+      }
+      expect(c1.bitLength, 7);
+      expect(c2.bitLength, 5);
+      expect(c3.bitLength, 11);
     });
   });
 }