+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- `CompositeSet.clone` returns `CompositeSet`.
+- Support `shiftLeft` in many operations.
+- Support multiply and clone of counters.
+- Use `package:pedantic` and Dart 2.2 features.
+
 ## 1.0.1
 
 - `asUint64Iterable()` added to provide interface-level compatibility for future compressed bit arrays.